Revert "feat/CX-36-add-react-routing:Added-routing"
