--- src/pages/projectPagina.tsx
@@ -1,26 +1,0 @@
-/*
-  This program has been developed by students from the bachelor Computer Science at Utrecht
-  University within the Software Project course.
-
-  © Copyright Utrecht University (Department of Information and Computing Sciences)
-*/
-
-
-import { Link } from "react-router"
-
-/*
-this is the main entry point into the projects page. 
-It is currently empty.
-*/
-
-function ProjectPagina() 
-{
-    return (
-    <>
-        <h1>Please create the project page here!</h1>
-        <Link to = "/"> go to home </Link>
-    </>
-    )
-}
-
-export default ProjectPagina
