--- src/pages/settings.tsx
@@ -1,25 +1,0 @@
-/*
-  This program has been developed by students from the bachelor Computer Science at Utrecht
-  University within the Software Project course.
-
-  © Copyright Utrecht University (Department of Information and Computing Sciences)
-*/
-
-import { Link } from "react-router"
-
-/*
-This is the main entrypoint into the settings page.
-It is currently empty.
-*/
-
-function SettingsPagina() 
-{
-    return (
-    <>
-        <h1>Please create the settings page here!</h1>
-        <Link to = "/"> go to home </Link>
-    </>
-    )
-}
-
-export default SettingsPagina
